requirements.txt
# Core dependencies
pydantic>=2.5.2
PyYAML>=6.0.1
openai>=1.12.0
python-docx>=0.8.11

# Rich terminal interface
rich>=13.7.0

# Type hints support
typing-extensions>=4.9.0

# Development tools
pytest>=7.4.3
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0

# Optional: for better YAML handling
ruamel.yaml>=0.18.5

# Optional: for better type checking
types-PyYAML>=6.0.12.12
types-python-docx>=0.8.11.20240106